plant_ai/requirements.txt
flask>=3.0.0,<4.0.0
numpy>=1.26.0,<3.0.0
pillow>=10.0.0,<12.0.0
tensorflow>=2.16.0,<3.0.0
